+++ src/tests/agenda/views/test_agenda_schedule.py
@@ -110,6 +110,37 @@ def test_schedule_page_text_table(
 
 
 @pytest.mark.django_db
+def test_schedule_page_text_table_explicit_header(
+    client, django_assert_num_queries, event, speaker, slot, schedule, other_slot,
+):
+    url = event.urls.schedule
+    with django_assert_num_queries(18):
+        response = client.get(url, follow=True, HTTP_ACCEPT='text/plain')
+    assert response.status_code == 200
+    title_lines = textwrap.wrap(slot.submission.title, width=16)
+    content = response.content.decode()
+    for line in title_lines:
+        assert line in content
+
+
+@pytest.mark.parametrize('header,target', (
+    ('application/json', 'frab_json'),
+    ('application/xml', 'frab_xml'),
+))
+@pytest.mark.django_db
+def test_schedule_page_redirects(
+    client, django_assert_num_queries, event, speaker, slot, schedule, other_slot,
+    header, target
+):
+    url = event.urls.schedule
+    with django_assert_num_queries(16):
+        response = client.get(url, HTTP_ACCEPT=header)
+    assert response.status_code == 303
+    assert response._headers['location'][1] == getattr(event.urls, target).full()
+    assert response.content.decode() == ''
+
+
+@pytest.mark.django_db
 def test_schedule_page_text_list(
     client, django_assert_num_queries, event, speaker, slot, schedule, other_slot
 ):